+++ build_tools/aws-sdk-code-generator/spec/protocols_spec.rb
@@ -143,6 +143,14 @@ def match_req_headers(group, test_case, http_req, it)
   end
 end
 
+def exclude_req_headers(group, test_case, http_req, it)
+  if excluded_headers = test_case['serialized']['excludeHeaders']
+    headers = normalize_headers(http_req.headers)
+    excluded_headers = normalize_headers(excluded_headers)
+    it.expect(headers).to_not include(excluded_headers)
+  end
+end
+
 def match_req_body(group, suite, test_case, http_req, it)
   protocol = suite['metadata']['protocol']
   if expected_body = test_case['serialized']['body']
@@ -199,6 +207,7 @@ fixtures.each do |directory, files|
           match_req_method(group, test_case, ctx.http_request, self)
           match_req_uri(group, test_case, ctx.http_request, self)
           match_req_headers(group, test_case, ctx.http_request, self)
+          exclude_req_headers(group, test_case, ctx.http_request, self)
           match_req_body(group, suite, test_case, ctx.http_request, self)
 
         end